src/app/requirements.txt
anyascii==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
contractions==0.0.52
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.6.0
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lime==0.2.0.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.0
Pillow==8.3.0
pyahocorasick==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.7.1
requests==2.25.1
requests-file==1.5.1
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
textsearch==0.0.21
threadpoolctl==2.1.0
tifffile==2021.7.2
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
